==> 7.embedding_bert.ipynb <==
# !pip install jieba

import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import gc
import time
import os

from random import randint

import numpy as np
import torch
import pandas as pd
import re
import os
os.environ["CUDA_VISIBLE_DEVICES"] ='0'

pub_info = pd.read_pickle('./pkl/pub_info.pkl')

pub_info.head()

### prepare

pub_info.isnull().sum()

pub_info['keywords'] = pub_info['keywords'].apply(lambda x: [] if type(x) == float else x)

# max_len=256
# def func(s):
#     s = re.sub(r'[^\w\s]', ' ', s)
#     wds = s.split()[:max_len]
#     return ' '.join(wds)

# cols = ['abstract', 'title', 'venue']
# for c in cols:
#     print(c)
#     pub_info[c] = pub_info[c].apply(func)

def func2(l):
    return ', '.join(l)

cols = ['authors', 'keywords', 'orgs']
for c in cols:
    print(c)
    pub_info[c] = pub_info[c].apply(func2)

pub_info.head()

### embedding

# !unzip ../../../multi/bert-large-nli-mean-tokens.zip -d ./bert-large-nli-mean-tokens/

import torch
torch.cuda.is_available()

from sentence_transformers import SentenceTransformer
model = SentenceTransformer('./bert-large-nli-mean-tokens')

# model = model.cuda()

pub_info_embed = pd.DataFrame()

cols = ['abstract', 'authors', 'keywords', 'title', 'venue', 'orgs']
for c in cols:
    print(c)
    pub_info_c = list(pub_info[c].values)
    pub_info_c_embed = model.encode(pub_info_c, show_progress_bar=True, batch_size=128)
    pub_info_embed[c] = pub_info_c_embed
    display(pub_info_embed.head(2))

pub_info_embed.to_pickle('./pkl/pub_info_embed_bert.pkl')
