==> 8.embedding2_bert.ipynb <==
import pandas as pd
import os

pub_info_embed = pd.read_pickle('./pkl/pub_info_embed_bert.pkl')
pub_ids = pd.read_pickle('./pkl/pub_info.pkl')[['paper_id']].reset_index(drop=True)

pub_ids.to_pickle('./embed/pub_ids.pkl')

embed_dir = './embed_bert/'
if not os.path.exists(embed_dir):
    os.mkdir(embed_dir)

pub_info_embed.head()

# seperate six col to save mem

cols = ['abstract', 'authors', 'keywords', 'title', 'venue', 'orgs']
for c in cols:
    print(c)
    pd.concat([pub_ids, pub_info_embed[[c]]], axis=1).to_pickle('./embed_bert/pub_info_%s.pkl' % c)
