==> 10.gen_feat_v2_bert.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import gc
import time
import os

feat_dir = './feat/'
if not os.path.exists(feat_dir):
    os.mkdir(feat_dir)

train_data = pd.read_pickle('./pkl/train_data.pkl')
test_data = pd.read_pickle('./pkl/valid_data.pkl')

data = pd.concat([train_data, test_data]).reset_index(drop=True)

def pidx(p, ps):
    ans = np.nan
    for i, p2 in enumerate(ps):
        if p == p2:
            ans = i
            break
    return ans

from scipy import spatial
def cos_sim(a, b):
    return 1 - spatial.distance.cosine(a, b)

from tqdm import tqdm_notebook

pub_col = ['abstract', 'keywords', 'title', 'venue', 'authors', 'orgs']
for pc in pub_col:
    print(pc)
    print('load data...')
    pub_info = pd.read_pickle('./embed_bert/pub_info_%s.pkl' % pc)
    author_pub_detail = pd.read_pickle('./embed_bert/author_pub_%s.pkl' % pc)
    pub_info.columns = ['paper_id', pc+'_a']
    author_pub_detail.columns = ['author_id', 'paper_ids', pc+'_b']
    data_c = data.merge(pub_info, 'left', 'paper_id').merge(author_pub_detail, 'left', 'author_id')
    print('find idx...')
    data_c['idx'] = data_c.apply(lambda row: pidx(row['paper_id'], row['paper_ids']) if row['label'] == 1 else np.nan, axis=1)
    display(data_c.head(3))
    print('del...')
    for i in tqdm_notebook(range(len(data_c))):
        if pd.isna(data_c.loc[i, 'idx']):
            continue
        v = list(data_c.loc[i, pc+'_b'])
        del v[data_c.loc[i, 'idx'].astype(int)]
        data_c.set_value(i, pc+'_b', v)
    data_c = data_c[[pc+'_a', pc+'_b']]
    display(data_c.head(3))
    print('cal sim...')
    sims = []
    for ca, cb in tqdm_notebook(data_c.values):
        ss = []
        for b in cb:
            cs = cos_sim(ca, b)
            ss.append(cs)
        if len(ss) == 0:
            ss.append(0)
        sims.append(ss)
#     sims_df = pd.DataFrame(sims, columns=[pc+'_sims'])
    sims_df = pd.DataFrame()
    sims_df[pc+'_sims'] = sims
    display(sims_df.head(3))
    print('yi ba suo ha...')
    sims_df[pc+'_sims_'+'min'] = sims_df[pc+'_sims'].apply(np.min)
    sims_df[pc+'_sims_'+'max'] = sims_df[pc+'_sims'].apply(np.max)
    sims_df[pc+'_sims_'+'mean'] = sims_df[pc+'_sims'].apply(np.mean)
    sims_df[pc+'_sims_'+'std'] = sims_df[pc+'_sims'].apply(np.std)
    sims_df[pc+'_sims_'+'mm2'] = (sims_df[pc+'_sims_'+'min'] + sims_df[pc+'_sims_'+'max']) / 2
    display(sims_df.head(3))
    sims_df[[pc+'_sims_'+'min', pc+'_sims_'+'max', pc+'_sims_'+'mean', pc+'_sims_'+'std', pc+'_sims_'+'mm2']].to_pickle('./feat/%s_sims_st_bert.pkl' % (pc))
    del pub_info, author_pub_detail, data_c, sims, sims_df
    gc.collect()
