==> 11.gen_feat_v3_bert.ipynb <==
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
import gc
import time
import os

feat_dir = './feat/'
if not os.path.exists(feat_dir):
    os.mkdir(feat_dir)

train_data = pd.read_pickle('./pkl/train_data.pkl')
test_data = pd.read_pickle('./pkl/valid_data.pkl')

data = pd.concat([train_data, test_data]).reset_index(drop=True)
data.head()

train_author_paper_ids = pd.read_pickle('./pkl/train_author_paper_ids.pkl')
whole_author_name_paper_ids = pd.read_pickle('./pkl/whole_author_name_paper_ids.pkl')

author_pub_ids = whole_author_name_paper_ids[['author_id','paper_ids']].merge(train_author_paper_ids, 'left', 'author_id')

author_pub_ids['paper_ids_x_len'] = author_pub_ids['paper_ids_x'].apply(len)
author_pub_ids['paper_ids_y_len'] = author_pub_ids['paper_ids_y'].apply(lambda x: 0 if type(x) == float else len(x))

author_pub_ids['paper_ids'] = author_pub_ids.apply(lambda row: list(set(row['paper_ids_x']) | (set() if type(row['paper_ids_y']) == float else set(row['paper_ids_y']))), axis=1)

author_pub_ids['paper_ids_len'] = author_pub_ids['paper_ids'].apply(len)

author_pub_ids.drop(columns=['paper_ids_x', 'paper_ids_y', 'paper_ids_x_len', 'paper_ids_y_len'], inplace=True)

author_pub_ids.head()

author_pub_ids.shape

data['author_org'].nunique()

author_id_org_map = pd.read_pickle('./pkl/author_id_org_map.pkl')

author_id_org_map.head()

author_id_org_map['orgs'].apply(len).sum()

# org embedding
from collections import OrderedDict
org_embedding = OrderedDict()
for o in data['author_org']:
    if o == '' or pd.isna(o):
            continue
    if type(o) == float:
        print('err %f' % o)
    org_embedding[o] = np.nan

for os in author_id_org_map['orgs']:
    for o in os:
        if o == '' or pd.isna(o):
            continue
        if type(o) == float:
            print('err %f' % o)
        org_embedding[o] = np.nan

len(org_embedding)

orgs = list(org_embedding.keys())

orgs[:5]

from random import randint

import numpy as np
import torch
import pandas as pd
import re
import os
os.environ["CUDA_VISIBLE_DEVICES"] ='0'

# max_len=256
# def func(s):
#     s = s.lower()
#     s = re.sub(r'[^\w\s]', ' ', s)
#     wds = s.split()[:max_len]
#     return ' '.join(wds)

# orgs = list(map(func, orgs))
# orgs[:5]

from sentence_transformers import SentenceTransformer
model = SentenceTransformer('./bert-large-nli-mean-tokens')

# org_embed = model.encode(orgs, bsize=32, tokenize=False, verbose=True)
org_embed = model.encode(orgs, show_progress_bar=True, batch_size=128)

print(len(org_embed))
print(org_embed[0].shape)

for i, k in enumerate(org_embedding.keys()):
    org_embedding[k] = org_embed[i]

data = data.merge(author_pub_ids, 'left', 'author_id').merge(author_id_org_map, 'left', 'author_id')

data.head()

def pidx(p, ps):
    ans = np.nan
    for i, p2 in enumerate(ps):
        if p == p2:
            ans = i
            break
    return ans
data['idx'] = data.apply(lambda row: pidx(row['paper_id'], row['paper_ids']) if row['label'] == 1 else np.nan, axis=1)
display(data.head(3))

print('del...')
from tqdm import tqdm_notebook
for i in tqdm_notebook(range(len(data))):
    if pd.isna(data.loc[i, 'idx']):
        continue
    v = list(data.loc[i, 'orgs'])
    del v[data.loc[i, 'idx'].astype(int)]
    data.set_value(i, 'orgs', v)

data.head()

data['orgs'] = data['orgs'].apply(set).apply(list)

from scipy import spatial
def cos_sim(a, b):
    return 1 - spatial.distance.cosine(a, b)

from tqdm import tqdm_notebook
org_sims = []
for author_org, orgs in tqdm_notebook(data[['author_org', 'orgs']].values):
    sim = []
    if author_org == '' or pd.isna(author_org):
        pass
    else:
        for o in orgs:
            if o == '' or pd.isna(o):
                continue
            s = cos_sim(org_embedding[author_org], org_embedding[o])
            sim.append(s)
    if len(sim) == 0:
        sim = [0]
    org_sims.append(sim)

data['org_sims'] = org_sims
data.head()

data['orgs_sims_max2'] = data['org_sims'].apply(np.max)
data['orgs_sims_min2'] = data['org_sims'].apply(np.min)
data['orgs_sims_mean2'] = data['org_sims'].apply(np.mean)
data['orgs_sims_std2'] = data['org_sims'].apply(np.std)

data.head(3)

data['paper_num'] = data['paper_ids_len'] - (data['label'] == 1).astype(int)

data['orgs_num'] = data['orgs'].apply(len) + (data['label'] == 1).astype(int)

data[['paper_num', 'orgs_num', 'orgs_sims_max2', 'orgs_sims_min2', 'orgs_sims_mean2', 'orgs_sims_std2']].to_pickle('./feat/feat_v3_bert.pkl')
